+++ enum_name_map_arg_bloc/lib/src/saut.dart
@@ -631,7 +631,7 @@ class Saut {
 
   /// Calls [back] repeatedly until found the page.
   static void backToPage(BuildContext context, Enum page) =>
-      backToPageName(context, page.name);
+      backToPageName(context, effectiveRouteNameBuilder(page));
 
   /// Trick explained here: https://github.com/flutter/flutter/issues/20451
   /// Note `ModalRoute.of(context).settings.name` doesn't always work.